+++ sdk/core/azure-core/src/main/java/com/azure/core/http/HttpClient.java
@@ -4,6 +4,7 @@
 package com.azure.core.http;
 
 import com.azure.core.implementation.http.HttpClientProviders;
+import com.azure.core.util.Context;
 import reactor.core.publisher.Mono;
 
 /**
@@ -11,12 +12,29 @@ import reactor.core.publisher.Mono;
  */
 public interface HttpClient {
     /**
+     * Key for {@link Context} where the value is a boolean flag that indicates whether the {@link HttpResponse} body
+     * should be eagerly read and buffered into memory.
+     */
+    String EAGERLY_READ_RESPONSE_CONTEXT_KEY = "azure-eagerly-read-response";
+
+    /**
      * Send the provided request asynchronously.
      *
      * @param request The HTTP request to send.
-     * @return A {@link Mono} that emits response asynchronously.
+     * @return A {@link Mono} that emits the response asynchronously.
      */
     Mono<HttpResponse> send(HttpRequest request);
+
+    /**
+     * Sends the provided request asynchronously with contextual information.
+     *
+     * @param request The HTTP request to send.
+     * @param context Contextual information about the request.
+     * @return A {@link Mono} that emits the response asynchronously.
+     */
+    default Mono<HttpResponse> send(HttpRequest request, Context context) {
+        return send(request);
+    }
 
     /**
      * Create default {@link HttpClient} instance.