+++ sdk/core/azure-core-http-netty/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.7.0-beta.1 (Unreleased)
 
+### New Features
+
+- Added functionality to eagerly read HTTP response bodies into memory when they will be deserialized into a POJO.
+
 ### Bug Fixes
 
 - Fixed a bug where a connection would remain active when timed out instead of being closed.